refactor(encrypt-submission): combines validation together with handler
