+++ src/app/modules/form/admin-form/admin-form.routes.ts
@@ -430,18 +430,6 @@ AdminFormsRouter.get(
 AdminFormsRouter.get(
   '/:formId([a-fA-F0-9]{24})/adminform/submissions/download',
   withUserAuthentication,
-  celebrate({
-    [Segments.QUERY]: Joi.object()
-      .keys({
-        startDate: Joi.date().format('YYYY-MM-DD').raw(),
-        endDate: Joi.date()
-          .format('YYYY-MM-DD')
-          .greater(Joi.ref('startDate'))
-          .raw(),
-        downloadAttachments: Joi.boolean().default(false),
-      })
-      .and('startDate', 'endDate'),
-  }),
   EncryptSubmissionController.handleStreamEncryptedResponses,
 )
